Add notifier interface
This adds a notifier interface that can be used by a storage driver. The
intent is to dispatch CRUD events to different backends and through
different means. This could be a message-queue, a long polling session,
Server-Sent Events or anything else.

The main idea is to keep the storage and CRUD operations separated from
the eventing needs apps will have.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>