Add controller interface for directories
This creates constructs for applications to simply instantiate a
controller which will ensure that directories are up-to-date and
persistent in a local database. It also listens for events from an
event-queue and calls the reconcile function respectively.

The reconcile function is left as an interface, as reconciliation is
application specific. e.g. we don't know what users will do with a
directory change, but we want to make sure they know what to do.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>