treemanager: Expose read-only mode and fast reads
This exposes the driver's read-only mode and fast reads options to the
tree manager server. When used, this will run the tree manager in a mode
that disallows any write operations. If fast reads are used, it will
attempt to use the fast read or inconsistent mode. In CockroachDB this
translates to using follower reads to allow an instance to fetch data
from a follower instead of relying on the leader.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>