Test refactoring
This adds more unit test-style testing for the tree manager. e.g. we now
focus on testing the tree manager itself instead of doing so with all
the moving parts.

This required us to make some common tests we could share between this
and the integration tests, so that was done too.

This also changed a bunch of interfaces in order to better expose the
functionality of the tree manager.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>