tree manager: Use auditevent library to provide audit functionality
This uses the [auditevent](https://github.com/metal-toolbox/auditevent)
library to add audit logs to the tree manager. So, it is now expected
that the tree manager will run with a sidecar container that outputs the
audit logs.

Relevant tests were added, as well as a container in the compose file to
ease development. Note that the logs will be in `.audit/audit.log`.

Note that it REALLY is necessary for the sidecar container to be
listening, else the main function will hang indefinitely. This is
inteded, as we don't want to start the server unless audit logs are
in place.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>