app: Make HTTP client optional
This makes the HTTP client an optional parameter for the application
Reconciler.

In order to ensure that the tree structure in the application doesn't
drift, this provides a relevant interface (`Seeder`) to allow service
authors to trigger full reconciliation if needed without having to rely
on a built-in client for their application.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>